# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: AI-Assisted Support Ticket Platform API
  version: 1.0.0
  description: External API contract. Internal service APIs are documented in service detail designs.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
...
paths:
...
  /v1/tickets:
    post:
      tags: [Tickets]
      operationId: createTicket
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: {type: string, format: uuid}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateTicketRequest'}
      responses:
        '201':
          description: Created; AI analysis is asynchronous
          headers:
            ETag:
              schema: {type: string, example: '"0"'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Ticket'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '409': {$ref: '#/components/responses/Conflict'}
    get:
      tags: [Tickets]
      operationId: listTickets
      parameters:
        - {name: status, in: query, schema: {$ref: '#/components/schemas/TicketStatus'}}
        - {name: priority, in: query, schema: {$ref: '#/components/schemas/Priority'}}
        - {name: category, in: query, schema: {$ref: '#/components/schemas/Category'}}
        - {name: assigneeId, in: query, schema: {type: string, format: uuid}}
        - {name: requesterId, in: query, schema: {type: string, format: uuid}}
        - {name: createdFrom, in: query, schema: {type: string, format: date-time}}
        - {name: createdTo, in: query, schema: {type: string, format: date-time}}
        - {name: q, in: query, schema: {type: string, maxLength: 200}}
        - {name: page, in: query, schema: {type: integer, minimum: 0, default: 0}}
        - {name: size, in: query, schema: {type: integer, minimum: 1, maximum: 50, default: 20}}
        - name: sort
          in: query
          schema:
            type: string
            default: updatedAt,desc
            example: createdAt,asc
      responses:
        '200':
          description: Page of authorized tickets
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TicketPage'}
  /v1/tickets/{ticketId}:
    parameters:
      - $ref: '#/components/parameters/TicketId'
    get:
      tags: [Tickets]
      operationId: getTicket
      responses:
        '200':
          description: Ticket detail
          headers:
            ETag:
              schema: {type: string, example: '"4"'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Ticket'}
        '404': {$ref: '#/components/responses/NotFound'}
    patch:
      tags: [Tickets]
      operationId: updateTicket
      parameters:
        - name: If-Match
          in: header
          required: true
          schema: {type: string, example: '"4"'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UpdateTicketRequest'}
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Ticket'}
        '409': {$ref: '#/components/responses/Conflict'}
        '412': {$ref: '#/components/responses/PreconditionFailed'}
        '428': {$ref: '#/components/responses/PreconditionRequired'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TicketId:
      name: ticketId
      in: path
      required: true
      schema: {type: string, format: uuid}
  schemas:
...
    CreateTicketRequest:
      type: object
      required: [subject, description]
      properties:
        subject: {type: string, minLength: 3, maxLength: 300}
        description: {type: string, minLength: 10, maxLength: 20000}
        priority: {$ref: '#/components/schemas/Priority'}
    UpdateTicketRequest:
      type: object
      minProperties: 1
      properties:
        subject: {type: string, minLength: 3, maxLength: 300}
        description: {type: string, minLength: 10, maxLength: 20000}
        status: {$ref: '#/components/schemas/TicketStatus'}
        priority: {$ref: '#/components/schemas/Priority'}
        category: {$ref: '#/components/schemas/Category'}
        assigneeId: {type: string, format: uuid, nullable: true}
    Ticket:
      type: object
      required: [id, number, subject, description, priority, status, version, createdAt, updatedAt]
      properties:
        id: {type: string, format: uuid}
        number: {type: string, example: SUP-00001234}
        subject: {type: string}
        description: {type: string}
        category: {$ref: '#/components/schemas/Category'}
        priority: {$ref: '#/components/schemas/Priority'}
        status: {$ref: '#/components/schemas/TicketStatus'}
        requesterId: {type: string, format: uuid}
        assigneeId: {type: string, format: uuid, nullable: true}
        aiAnalysis: {$ref: '#/components/schemas/AiAnalysis'}
        version: {type: integer, format: int64}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    AiAnalysis:
      type: object
      properties:
        status: {type: string, enum: [QUEUED, RUNNING, SUCCEEDED, FAILED]}
        summary: {type: string, nullable: true}
        suggestedCategory: {$ref: '#/components/schemas/Category'}
        suggestedPriority: {$ref: '#/components/schemas/Priority'}
        generatedAt: {type: string, format: date-time, nullable: true}
...
    TicketPage:
      type: object
      required: [content, page, size, totalElements, totalPages]
      properties:
        content: {type: array, items: {$ref: '#/components/schemas/Ticket'}}
        page: {type: integer}
        size: {type: integer}
        totalElements: {type: integer, format: int64}
        totalPages: {type: integer}
    TicketStatus:
      type: string
      enum: [OPEN, IN_PROGRESS, WAITING_CUSTOMER, RESOLVED, CLOSED]
    Priority:
      type: string
      enum: [LOW, MEDIUM, HIGH, URGENT]
    Category:
      type: string
      enum: [ACCOUNT, PAYMENT, TECHNICAL, GENERAL, OTHER]
    Problem:
      type: object
      required: [type, title, status, code, correlationId]
      properties:
        type: {type: string, format: uri-reference}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        instance: {type: string, format: uri-reference}
        code: {type: string}
        correlationId: {type: string}
        violations:
          type: array
          items:
            type: object
            properties:
              field: {type: string}
              message: {type: string}
  responses:
    BadRequest:
      description: Invalid request
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}
...
    NotFound:
      description: Resource not found
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}
    Conflict:
      description: Domain conflict
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}
    PreconditionFailed:
      description: Optimistic-lock version conflict
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}
    PreconditionRequired:
      description: If-Match is required
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}